// src/app/budget/types/budget.d.ts
interface FilterOptions {
  department: string; // Department ID as string
  departmentname: string; // Department name as string
  year: string; // Year as string
  subdepartmentId: number;
  subdepartmentName: string;
}
interface BudgetFilterFormProps {
  filters: FilterOptions;
  handleSelect: (
    name: string,
    value: number | string | { id: number; departmentname: string },
  ) => void;
  budgetId: number | null;
  setBugetId: (val: number | null) => void;
  status: string | undefined;
  setStatus: (val: string) => void;
}